+++ core/genesis_alloc.go
@@ -7,13 +7,18 @@ type genesisAlloc map[common.Address]GenesisAccount
 type GenesisAccount struct {
 	Address common.Address `json:"address" yaml:"address"`
 	Balance uint64         `json:"balance" yaml:"balance"`
-	Auth    string         `json:"auth" yaml:"auth"`
 }
 
 func developerNetAlloc() genesisAlloc {
 	return map[common.Address]GenesisAccount{
 		common.HexToAddress("0x0000000000000000000000000000000000000000"): {
-			Balance: 100e9,
+			Balance: 1e12,
+		},
+		common.HexToAddress("0x0000000000000000000000000000000000000000"): {
+			Balance: 1e9,
+		},
+		common.HexToAddress("0x0000000000000000000000000000000000000000"): {
+			Balance: 1e6,
 		},
 	}
 }
@@ -26,11 +31,11 @@ func defaultMainNetAlloc() genesisAlloc {
 		common.HexToAddress("0x36527b4481018dff6d3400a2271d070910453420"): {
 			Balance: 110e9,
 		},
-		common.HexToAddress("0x3c0b3b41a1e027d3e759612af08844f1cca0dde3"): {
+		common.HexToAddress("0xF5f998c761F0CE7e2b15df323e6862D0C31c9F6F"): {
 			Balance: 110e9,
 		},
 		common.HexToAddress("0x40b2121f4eb40B6863A08D08C567CC1C995f971F"): {
-			Balance: 110e6,
+			Balance: 110e9,
 		},
 	}
 }