+++ core/types/block.go
@@ -2,6 +2,7 @@ package types
 
 import (
 	"bytes"
+	"crypto/sha256"
 	"encoding/gob"
 	"github.com/ethereum/go-ethereum/common"
 )
@@ -30,8 +31,84 @@ func (bh BlockHeader) Serialize() ([]byte, error) {
 	return buf.Bytes(), nil
 }
 
-// Deserialize deserializes a block header from gob encoding
+// Deserialize deserializes binary data to BlockHeader
 func (bh *BlockHeader) Deserialize(d []byte) error {
 	decoder := gob.NewDecoder(bytes.NewReader(d))
 	return decoder.Decode(bh)
+}
+
+// NewBlock creates and returns Block
+func NewBlock(header BlockHeader, txs []SignedTx, receipts []*Receipt) *Block {
+	return &Block{
+		BlockHeader:  header,
+		Transactions: txs,
+	}
+}
+
+// Block represents Blockchain state change interface
+type Block struct {
+	BlockHeader
+	Transactions []SignedTx `json:"transactions" yaml:"transactions"`
+
+	//size int64
+
+	//ReceivedAt int64 `json:"received_at" yaml:"received_at"`
+}
+
+// SetHash sets a hash of the block
+func (b *Block) SetHash() error {
+	enc, err := b.Serialize()
+	if err != nil {
+		return err
+	}
+
+	hash := sha256.Sum256(enc)
+	b.Hash = hash
+	return nil
+}
+
+// Size returns encoded block value byte length
+func (b *Block) Size() error {
+	enc, err := b.Serialize()
+	if err != nil {
+		return err
+	}
+
+	hash := sha256.Sum256(enc)
+	b.Hash = hash
+	return nil
+}
+
+// HashTransactions returns a hash of the transactions in the block
+func (b *Block) HashTransactions() ([]byte, error) {
+	var txHashes [][]byte
+	var txHash [32]byte
+
+	for _, tx := range b.Transactions {
+		hash, err := tx.Hash()
+		if err != nil {
+			return nil, err
+		}
+		txHashes = append(txHashes, hash)
+	}
+	txHash = sha256.Sum256(bytes.Join(txHashes, []byte{}))
+
+	return txHash[:], nil
+}
+
+// Serialize serializes the block
+func (b *Block) Serialize() ([]byte, error) {
+	var result bytes.Buffer
+	encoder := gob.NewEncoder(&result)
+	if err := encoder.Encode(b); err != nil {
+		return nil, err
+	}
+
+	return result.Bytes(), nil
+}
+
+// Deserialize deserializes binary data to block
+func (b *Block) Deserialize(d []byte) error {
+	decoder := gob.NewDecoder(bytes.NewReader(d))
+	return decoder.Decode(b)
 }